+++ frontend/techpick/src/features/Draggable/pick.css.ts
@@ -1,0 +1,18 @@
+import { commonThemeContract } from '@/shared/themes';
+import { style } from '@vanilla-extract/css';
+
+export const linkWrapper = style({
+  display: 'flex',
+  flexDirection: 'column',
+  justifyContent: 'center',
+  alignItems: 'center',
+  width: '100px',
+  height: '100px',
+  padding: `0 ${commonThemeContract.space.md}`,
+  fontWeight: 300,
+  textAlign: 'center',
+  borderRadius: '4px',
+  ':hover': {
+    backgroundColor: '#efefef',
+  },
+});
